Extracts command execution to its own package
Refactors downloader package name to download
Refactors DownloadFromS3 function name
